refactor: extract event to reactable api adapter
